+++ src/themes/five/src/components/OrderDetails/index.js
@@ -11,11 +11,6 @@ import {
   GoogleMapsMap
 } from 'ordering-components'
 import RiUser2Fill from '@meronex/icons/ri/RiUser2Fill'
-import BiStoreAlt from '@meronex/icons/bi/BiStoreAlt'
-import AiFillExclamationCircle from '@meronex/icons/ai/AiFillExclamationCircle'
-import BsPhone from '@meronex/icons/bs/BsPhone'
-import BiMessageRounded from '@meronex/icons/bi/BiMessageRounded'
-import BsInfoCircle from '@meronex/icons/bs/BsInfoCircle'
 
 import { Button } from '../../styles/Buttons'
 import { NotFoundSource } from '../NotFoundSource'
@@ -25,6 +20,9 @@ import { Modal } from '../Modal'
 import { Messages } from '../Messages'
 import { ReviewOrder } from '../../../../../components/ReviewOrder'
 import { ProductShare } from '../../../../../components/ProductShare'
+import { OrderBillSection } from './OrderBillSection'
+import { ActionsSection } from './ActionsSection'
+import { OrderPreferencesSection } from './OrderPreferencesSections'
 
 import {
   Container,
@@ -34,7 +32,6 @@ import {
   OrderBusiness,
   BusinessWrapper,
   BusinessInfo,
-  ActionsBlock,
   OrderInfo,
   StatusBar,
   SectionTitle,
@@ -44,12 +41,9 @@ import {
   OrderDriver,
   WrapperDriver,
   OrderProducts,
-  OrderBill,
   SkeletonBlockWrapp,
   SkeletonBlock,
   ShareOrder,
-  MessagesIcon,
-  ExclamationWrapper,
   WrapperLeftContainer,
   WrapperRightContainer,
   Divider,
@@ -57,8 +51,6 @@ import {
   ReviewOrderLink,
   SkeletonWrapper,
   ReviewWrapper,
-  Exclamation,
-  CommentContainer,
   NewOrder,
   OrderActions,
   TitleContainer,
@@ -69,7 +61,6 @@ import {
   HeaderTitle
 } from './styles'
 import { useTheme } from 'styled-components'
-import { verifyDecimals } from '../../../../../utils'
 import { ReviewProduct } from '../../../../../components/ReviewProduct'
 import { ReviewDriver } from '../../../../../components/ReviewDriver'
 import { TaxInformation } from '../TaxInformation'
@@ -95,7 +86,7 @@ const OrderDetailsUI = (props) => {
   const [{ configs }] = useConfig()
   const theme = useTheme()
   const [events] = useEvent()
-  const [{ parsePrice, parseNumber, parseDate }] = useUtils()
+  const [{ parsePrice, parseDate }] = useUtils()
   const [, { deleteUserCustomer }] = useCustomer()
   const [{ carts }, { refreshOrderOptions }] = useOrder()
 
@@ -109,16 +100,6 @@ const OrderDetailsUI = (props) => {
   const [openTaxModal, setOpenTaxModal] = useState({ open: false, tax: null })
 
   const { order, loading, businessData, error } = props.order
-  const placeSpotTypes = [3, 4]
-
-  const walletName = {
-    cash: {
-      name: t('PAY_WITH_CASH_WALLET', 'Pay with Cash Wallet')
-    },
-    credit_point: {
-      name: t('PAY_WITH_CREDITS_POINTS_WALLET', 'Pay with Credit Points Wallet')
-    }
-  }
 
   const getOrderStatus = (s) => {
     const status = parseInt(s)
@@ -156,16 +137,6 @@ const OrderDetailsUI = (props) => {
 
   const handleGoToPage = (data) => {
     events.emit('go_to_page', data)
-  }
-
-  const handleOpenMessages = (data) => {
-    setOpenMessages(data)
-    readMessages()
-    if (order?.unread_count > 0) {
-      data.business
-        ? setUnreadAlert({ ...unreadAlert, business: false })
-        : setUnreadAlert({ ...unreadAlert, driver: false })
-    }
   }
 
   const unreadMessages = () => {
@@ -213,20 +184,6 @@ const OrderDetailsUI = (props) => {
     }
   }
 
-  const getIncludedTaxes = () => {
-    if (order?.taxes?.length === 0) {
-      return order.tax_type === 1 ? order?.summary?.tax ?? 0 : 0
-    } else {
-      return order?.taxes.reduce((taxIncluded, tax) => {
-        return taxIncluded + (tax.type === 1 ? tax.summary?.tax : 0)
-      }, 0)
-    }
-  }
-
-  const getIncludedTaxesDiscounts = () => {
-    return order?.taxes?.filter(tax => tax?.type === 1)?.reduce((carry, tax) => carry + (tax?.summary?.tax_after_discount ?? tax?.summary?.tax), 0)
-  }
-
   const closeOrderModal = (e) => {
     const outsideModal = !window.document.getElementById('app-modals') ||
       !window.document.getElementById('app-modals').contains(e.target)
@@ -244,6 +201,16 @@ const OrderDetailsUI = (props) => {
     deleteUserCustomer(true)
     refreshOrderOptions()
     handleGoToPage({ page: 'home' })
+  }
+
+  const ActionsSectionProps = {
+    order,
+    handleBusinessRedirect,
+    businessData,
+    unreadAlert,
+    setOpenMessages,
+    readMessages,
+    setUnreadAlert
   }
 
   useEffect(() => {
@@ -294,55 +261,6 @@ const OrderDetailsUI = (props) => {
     )
   }
 
-  const ActionsSection = ({ actionType }) => { // driver or business
-    return (
-      <ActionsBlock>
-        {order?.[actionType] && order?.[actionType]?.phone &&
-          <span onClick={() => window.open(`tel:${order?.[actionType]?.phone}`)}>
-            <BsPhone />
-          </span>}
-        {actionType === 'business' && (
-          <span>
-            <BiStoreAlt onClick={() => handleBusinessRedirect(businessData?.slug)} />
-          </span>
-        )}
-        <MessagesIcon onClick={() => handleOpenMessages({ driver: actionType === 'driver', business: actionType === 'business' })}>
-          {order?.unread_count > 0 && unreadAlert?.[actionType] && (
-            <ExclamationWrapper>
-              <AiFillExclamationCircle />
-            </ExclamationWrapper>
-          )}
-          <BiMessageRounded />
-        </MessagesIcon>
-      </ActionsBlock>
-    )
-  }
-
-  const OrderPreferencesSection = () => {
-    return (
-      <>
-        {order?.delivery_type === 1 && (
-          <CommentContainer>
-            <h3>{t('DELIVERY_PREFERENCE', 'Delivery preference')}</h3>
-            <span>{order?.delivery_option?.name ? t(order?.delivery_option?.name.toUpperCase().replace(/\s/g, '_'), order?.delivery_option?.name) : t('EITHER_WAY', 'Either way')}</span>
-          </CommentContainer>
-        )}
-        {order?.comment && (
-          <CommentContainer>
-            <h3>{t('COMMENT', 'Comment')}</h3>
-            <span>{order?.comment}</span>
-          </CommentContainer>
-        )}
-        {placeSpotTypes.includes(order?.delivery_type) && order?.place && (
-          <CommentContainer>
-            <h3>{t('SPOT', 'Spot')}</h3>
-            <span>{order?.place?.name}</span>
-          </CommentContainer>
-        )}
-      </>
-    )
-  }
-
   const OrderHeaderInfoSection = () => {
     return (
       <HeaderInfo>
@@ -370,230 +288,6 @@ const OrderDetailsUI = (props) => {
     )
   }
 
-  const OrderBillSection = () => {
-    return (
-      <OrderBill isCustomerMode={isCustomerMode}>
-        <table>
-          <tbody>
-            <tr>
-              <td>{t('SUBTOTAL', theme?.defaultLanguages?.SUBTOTAL || 'Subtotal')}</td>
-              <td>
-                {parsePrice(((order?.summary?.subtotal ?? order?.subtotal) + getIncludedTaxes()))}
-              </td>
-            </tr>
-            {(order?.summary?.discount > 0 ?? order?.discount > 0) && order?.offers?.length === 0 && (
-              <tr>
-                {order?.offer_type === 1 ? (
-                  <td>
-                    {t('DISCOUNT', theme?.defaultLanguages?.DISCOUNT || 'Discount')}{' '}
-                    <span>{`(${verifyDecimals(order?.offer_rate, parsePrice)}%)`}</span>
-                  </td>
-                ) : (
-                  <td>{t('DISCOUNT', theme?.defaultLanguages?.DISCOUNT || 'Discount')}</td>
-                )}
-                <td>- {parsePrice(order?.summary?.discount ?? order?.discount)}</td>
-              </tr>
-            )}
-            {
-              order?.offers?.length > 0 && order?.offers?.filter(offer => offer?.target === 1)?.map(offer => (
-                <tr key={offer.id}>
-                  <td>
-                    {offer.name}
-                    {offer.rate_type === 1 && (
-                      <span>{`(${verifyDecimals(offer?.rate, parsePrice)}%)`}</span>
-                    )}
-                    <Exclamation onClick={() => setOpenTaxModal({ open: true, data: offer, type: 'offer_target_1' })}>
-                      <BsInfoCircle size='20' color={theme.colors.primary} />
-                    </Exclamation>
-                  </td>
-                  <td>
-                    - {parsePrice(offer?.summary?.discount)}
-                  </td>
-                </tr>
-              ))
-            }
-            <tr>
-              <td>
-                <Divider />
-              </td>
-              <td>
-                <Divider />
-              </td>
-            </tr>
-            {order?.summary?.subtotal_with_discount > 0 && order?.summary?.discount > 0 && order?.summary?.total >= 0 && (
-              <tr>
-                <td>{t('SUBTOTAL_WITH_DISCOUNT', 'Subtotal with discount')}</td>
-                {order?.tax_type === 1 ? (
-                  <td>{parsePrice((order?.summary?.subtotal_with_discount + getIncludedTaxesDiscounts() ?? 0))}</td>
-                ) : (
-                  <td>{parsePrice(order?.summary?.subtotal_with_discount ?? 0)}</td>
-                )}
-              </tr>
-            )}
-            {
-              order?.taxes?.length === 0 && order?.tax_type === 2 && (
-                <tr>
-                  <td>
-                    {t('TAX', 'Tax')}
-                    <span>{`(${verifyDecimals(order?.tax, parseNumber)}%)`}</span>
-                  </td>
-                  <td>{parsePrice(order?.summary?.tax ?? 0)}</td>
-                </tr>
-              )
-            }
-            {
-              order?.fees?.length === 0 && (
-                <tr>
-                  <td>
-                    {t('SERVICE_FEE', 'Service fee')}
-                    <span>{`(${verifyDecimals(order?.service_fee, parseNumber)}%)`}</span>
-                  </td>
-                  <td>{parsePrice(order?.summary?.service_fee ?? 0)}</td>
-                </tr>
-              )
-            }
-            {
-              order?.taxes?.length > 0 && order?.taxes?.filter(tax => tax?.type === 2 && tax?.rate !== 0).map(tax => (
-                <tr key={tax?.id}>
-                  <td>
-                    {tax?.name || t('INHERIT_FROM_BUSINESS', 'Inherit from business')}
-                    <span>{`(${verifyDecimals(tax?.rate, parseNumber)}%)`}</span>
-                    <Exclamation onClick={() => setOpenTaxModal({ open: true, data: tax, type: 'tax' })}>
-                      <BsInfoCircle size='20' color={theme.colors.primary} />
-                    </Exclamation>
-                  </td>
-                  <td>{parsePrice(tax?.summary?.tax_after_discount ?? tax?.summary?.tax ?? 0)}</td>
-                </tr>
-              ))
-            }
-            {
-              order?.fees?.length > 0 && order?.fees?.filter(fee => !(fee?.fixed === 0 && fee?.percentage === 0))?.map(fee => (
-                <tr key={fee.id}>
-                  <td>
-                    {fee?.name || t('INHERIT_FROM_BUSINESS', 'Inherit from business')}
-                    ({fee?.fixed > 0 && `${parsePrice(fee?.fixed)} + `}{fee.percentage}%)
-                    <Exclamation onClick={() => setOpenTaxModal({ open: true, data: fee, type: 'fee' })}>
-                      <BsInfoCircle size='20' color={theme.colors.primary} />
-                    </Exclamation>
-                  </td>
-                  <td>{parsePrice(fee?.summary?.fixed + (fee?.summary?.percentage_after_discount ?? fee?.summary?.percentage) ?? 0)}</td>
-                </tr>
-              ))
-            }
-            {
-              order?.offers?.length > 0 && order?.offers?.filter(offer => offer?.target === 3)?.map(offer => (
-                <tr key={offer.id}>
-                  <td>
-                    {offer.name}
-                    {offer.rate_type === 1 && (
-                      <span>{`(${verifyDecimals(offer?.rate, parsePrice)}%)`}</span>
-                    )}
-                    <Exclamation onClick={() => setOpenTaxModal({ open: true, data: offer, type: 'offer_target_3' })}>
-                      <BsInfoCircle size='20' color={theme.colors.primary} />
-                    </Exclamation>
-                  </td>
-                  <td>
-                    - {parsePrice(offer?.summary?.discount)}
-                  </td>
-                </tr>
-              ))
-            }
-            {order?.summary?.delivery_price > 0 && (
-              <tr>
-                <td>{t('DELIVERY_FEE', theme?.defaultLanguages?.DELIVERY_FEE || 'Delivery Fee')}</td>
-                <td>{parsePrice(order?.summary?.delivery_price)}</td>
-              </tr>
-            )}
-            {
-              order?.offers?.length > 0 && order?.offers?.filter(offer => offer?.target === 2)?.map(offer => (
-                <tr key={offer.id}>
-                  <td>
-                    {offer.name}
-                    {offer.rate_type === 1 && (
-                      <span>{`(${verifyDecimals(offer?.rate, parsePrice)}%)`}</span>
-                    )}
-                    <Exclamation onClick={() => setOpenTaxModal({ open: true, data: offer, type: 'offer_target_2' })}>
-                      <BsInfoCircle size='20' color={theme.colors.primary} />
-                    </Exclamation>
-                  </td>
-                  <td>
-                    - {parsePrice(offer?.summary?.discount)}
-                  </td>
-                </tr>
-              ))
-            }
-            {(order?.summary?.driver_tip > 0 || order?.driver_tip > 0) && (
-              <tr>
-                <td>
-                  {t('DRIVER_TIP', theme?.defaultLanguages?.DRIVER_TIP || 'Driver tip')}{' '}
-                  {(order?.summary?.driver_tip > 0 || order?.driver_tip > 0) &&
-                    parseInt(configs?.driver_tip_type?.value, 10) === 2 &&
-                    !parseInt(configs?.driver_tip_use_custom?.value, 10) &&
-                    (
-                      <span>{`(${verifyDecimals(order?.driver_tip, parseNumber)}%)`}</span>
-                    )}
-                </td>
-                <td>{parsePrice(order?.summary?.driver_tip ?? order?.totalDriverTip)}</td>
-              </tr>
-            )}
-          </tbody>
-        </table>
-        <table className='total'>
-          <tbody>
-            <tr>
-              <td>{t('TOTAL', theme?.defaultLanguages?.TOTAL || 'Total')}</td>
-              <td>{parsePrice(order?.summary?.total ?? order?.total)}</td>
-            </tr>
-          </tbody>
-        </table>
-        {order?.payment_events?.length > 0 && (
-          <div style={{ marginTop: 10 }}>
-            <span style={{ fontSize: 20 }}>{t('PAYMENTS', 'Payments')}</span>
-            <div
-              style={{
-                width: '100%',
-                marginTop: 10
-              }}
-            >
-              {order?.payment_events?.map(event => event.amount > 0 && (
-                <div
-                  key={event.id}
-                  style={{
-                    display: 'flex',
-                    flexDirection: 'row',
-                    justifyContent: 'space-between',
-                    alignItems: 'center',
-                    marginBottom: 10
-                  }}
-                >
-                  <div
-                    style={{
-                      display: 'flex',
-                      flexDirection: 'column'
-                    }}
-                  >
-                    <span>
-                      {event?.wallet_event
-                        ? walletName[event?.wallet_event?.wallet?.type]?.name
-                        : event?.paymethod?.name}
-                    </span>
-                    {event?.data?.charge_id && (
-                      <span>
-                        {`${t('CODE', 'Code')}: ${event?.data?.charge_id}`}
-                      </span>
-                    )}
-                  </div>
-                  <span>
-                    -{parsePrice(event.amount)}
-                  </span>
-                </div>
-              ))}
-            </div>
-          </div>
-        )}
-      </OrderBill>
-    )
-  }
   return (
     <>
       {props.beforeElements?.map((BeforeElement, i) => (
@@ -707,7 +401,10 @@ const OrderDetailsUI = (props) => {
                       <>
                         <BusinessTitle>
                           <h2>{order?.business?.name}</h2>
-                          <ActionsSection actionType='business' />
+                          <ActionsSection
+                            {...ActionsSectionProps}
+                            actionType='business'
+                          />
                         </BusinessTitle>
                         <p>{order?.business?.email}</p>
                         <p>{order?.business?.cellphone}</p>
@@ -727,7 +424,10 @@ const OrderDetailsUI = (props) => {
                 {isCustomerMode ? (
                   <OrderMapSection />
                 ) : (
-                  <ActionsSection actionType='business' />
+                  <ActionsSection
+                    {...ActionsSectionProps}
+                    actionType='business'
+                  />
                 )}
               </OrderBusiness>
               <Divider />
@@ -750,7 +450,10 @@ const OrderDetailsUI = (props) => {
                       <OrderDriver isCustomerMode={isCustomerMode}>
                         <SectionTitleContainer>
                           <h2>{t('DRIVER', theme?.defaultLanguages?.DRIVER || 'Driver')}</h2>
-                          <ActionsSection actionType='driver' />
+                          <ActionsSection
+                            {...ActionsSectionProps}
+                            actionType='driver'
+                          />
                         </SectionTitleContainer>
                         <WrapperDriver>
                           <div className='photo'>
@@ -786,7 +489,10 @@ const OrderDetailsUI = (props) => {
                           </div>
                           <p>{order?.driver?.name} {order?.driver?.lastname}</p>
                         </WrapperDriver>
-                        <ActionsSection actionType='driver' />
+                        <ActionsSection
+                          {...ActionsSectionProps}
+                          actionType='driver'
+                        />
                       </OrderDriver>
                     </>
                   )}
@@ -819,12 +525,19 @@ const OrderDetailsUI = (props) => {
                   />
                 ))}
                 {isCustomerMode && (
-                  <OrderBillSection />
+                  <OrderBillSection
+                    order={order}
+                    isCustomerMode={isCustomerMode}
+                    setOpenTaxModal={setOpenTaxModal}
+                  />
                 )}
               </OrderProducts>
               {!isCustomerMode && (
                 <>
-                  <OrderBillSection />
+                  <OrderBillSection
+                    order={order}
+                    setOpenTaxModal={setOpenTaxModal}
+                  />
                   <Content className='order-content'>
                     {parseInt(configs?.guest_uuid_access?.value, 10) && order?.hash_key && (
                       <ShareOrder>
